+++ GenICam.Tests/Helper.cs
@@ -3,38 +3,46 @@ using Xunit;
 using GenICam;
 using GigeVision.Core;
 using GigeVision.Core.Models;
+using System.Collections.Generic;
+using org.mariuszgromada.math.mxparser;
+using System.Linq;
+using System.IO;
+using System;
 
 namespace GenICam.Tests
 {
     public class Helper
     {
         [Fact]
-        public void XMLParser()
+        public async void ReadAllRegisters()
         {
-            XmlDocument xmlDocument = new XmlDocument();
-            xmlDocument.Load("XMLSample_Breakfast.xml");
-            //xmlDocument.Load(filename: "Imperx.xml");
-
-
-            var xnsm = new XmlNamespaceManager(xmlDocument.NameTable);
-            var rootNode = xmlDocument["breakfast_menu"];
-            var uri = rootNode.Attributes["xmlns"].Value;
-            var ns = rootNode.Attributes["StandardNameSpace"].Value;
-            xnsm.AddNamespace(ns, uri);
-            var node = xmlDocument.SelectSingleNode($"//*[@Name='Sa']", xnsm);
-        }
-
-
-        [Fact]
-        public async void XMLHelper()
-        {
-
             XmlDocument xmlDocument = new XmlDocument();
             xmlDocument.Load("Imperx.xml");
             var genPort = new GenPort(new Gvcp());
             var xmlHelper = new XmlHelper(xmlDocument, genPort);
-             await xmlHelper.LoadUp();
+            await xmlHelper.LoadUp(true);
+
+            Assert.NotEmpty(xmlHelper.CategoryDictionary);
         }
 
+
+        [Theory]
+        [InlineData("(16=0)? 1: ( (0=1)?2:((0=2)? 3 :( (0=4)?4:((0=8)?5:((16=16)?6:((0=32)?7:8))))))")]
+        public void MathParser(string formula)
+        {
+            var expectedValue = 6; 
+            var actualValue = MathParserHelper.CalculateExpression(formula);
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData("(8))")]
+        [InlineData("((8)")]
+        public void GetBracket(string formula)
+        {
+            var expected = "(8)";
+            var actual  = MathParserHelper.GetBracketed(formula);
+            Assert.Equal(expected, actual);
+        }
     }
 }
