+++ GenICam/GenICam.csproj
@@ -1,10 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
     <Platforms>AnyCPU;x64</Platforms>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>2.1.1</Version>
+    <Version>2.1.3</Version>
 
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
   </PropertyGroup>
@@ -18,7 +18,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MathParser.org-mXparser" Version="5.0.3" />
+    <PackageReference Include="MathParser.org-mXparser" Version="5.0.7" />
     <PackageReference Include="Prism.Core" Version="8.1.97" />
   </ItemGroup>
 </Project>
