+++ GigeVision.Core/GigeVision.Core.csproj
@@ -41,7 +41,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="GenICam" Version="2.1.2" />
+    <PackageReference Include="GenICam" Version="2.1.3" />
     <PackageReference Include="Stira.WpfCore" Version="1.2.0" />
     <PackageReference Include="System.IO.Pipelines" Version="6.0.3" />
   </ItemGroup>
