+++ GigeVision.Core/Interfaces/IGvcp.cs
@@ -298,7 +298,10 @@ namespace GigeVision.Core.Interfaces
         /// </summary>
         /// <returns>Leave Status</returns>
         Task<bool> LeaveControl();
+        Task<(IPValue pValue, IRegister register)> GetRegister(string name);
+        Task<bool> ReadXmlFileAsync(string ip = null);
 
-        bool IsLoadingXml { get;}
+        bool IsLoadingXml { get; }
+        bool IsXmlFileLoaded { get; }
     }
 }
