+++ GigeVision.Core/Services/Camera.cs
@@ -1,9 +1,7 @@
-﻿using GenICam;
-using GigeVision.Core.Enums;
+﻿using GigeVision.Core.Enums;
 using GigeVision.Core.Interfaces;
 using Stira.WpfCore;
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace GigeVision.Core.Models
@@ -306,16 +304,21 @@ namespace GigeVision.Core.Models
             {
                 SetRxBuffer();
             }
-            if (Gvcp.RegistersDictionary.ContainsKey(nameof(RegisterName.AcquisitionStart)))
+            var acquisitionStart = (await Gvcp.GetRegister(nameof(RegisterName.AcquisitionStart))).pValue;
+            if (acquisitionStart != null)
             {
                 if (await Gvcp.TakeControl(true).ConfigureAwait(false))
                 {
                     SetupRxThread();
-                    if (((await Gvcp.RegistersDictionary[nameof(GvcpRegister.GevSCPHostPort)].pValue.SetValueAsync((uint)PortRx).ConfigureAwait(false)) as GvcpReply).Status == GvcpStatus.GEV_STATUS_SUCCESS)
+                    var gevSCPHostPort = (await Gvcp.GetRegister(nameof(GvcpRegister.GevSCPHostPort))).pValue;
+                    if ((await gevSCPHostPort.SetValueAsync((uint)PortRx).ConfigureAwait(false) as GvcpReply).Status == GvcpStatus.GEV_STATUS_SUCCESS)
                     {
-                        await Gvcp.RegistersDictionary[nameof(GvcpRegister.GevSCDA)].pValue.SetValueAsync(Converter.IpToNumber(ip2Send)).ConfigureAwait(false);
-                        await Gvcp.RegistersDictionary[nameof(GvcpRegister.GevSCPSPacketSize)].pValue.SetValueAsync(Payload).ConfigureAwait(false);
-                        if (((await Gvcp.RegistersDictionary[nameof(RegisterName.AcquisitionStart)].pValue.SetValueAsync(1).ConfigureAwait(false)) as GvcpReply).Status == GvcpStatus.GEV_STATUS_SUCCESS)
+                        var gevSCDA = (await Gvcp.GetRegister(nameof(GvcpRegister.GevSCDA))).pValue;
+                        await gevSCDA.SetValueAsync(Converter.IpToNumber(ip2Send)).ConfigureAwait(false);
+
+                        var gevSCPSPacketSize = (await Gvcp.GetRegister(nameof(GvcpRegister.GevSCPSPacketSize))).pValue;
+                        await gevSCPSPacketSize.SetValueAsync(Payload).ConfigureAwait(false);
+                        if (((await acquisitionStart.SetValueAsync(1).ConfigureAwait(false)) as GvcpReply).Status == GvcpStatus.GEV_STATUS_SUCCESS)
                         {
                             IsStreaming = true;
                         }
@@ -366,22 +369,22 @@ namespace GigeVision.Core.Models
             try
             {
                 await Gvcp.TakeControl().ConfigureAwait(false);
-                GvcpReply widthWriteReply = (await Gvcp.RegistersDictionary[nameof(RegisterName.Width)].pValue.SetValueAsync(width).ConfigureAwait(false)) as GvcpReply;
-                GvcpReply heightWriteReply = (await Gvcp.RegistersDictionary[nameof(RegisterName.Height)].pValue.SetValueAsync(width).ConfigureAwait(false)) as GvcpReply;
-
-                await Gvcp.RegistersDictionary[nameof(RegisterName.Height)].pValue.SetValueAsync(height).ConfigureAwait(false);
+                var widthPValue = (await Gvcp.GetRegister(nameof(RegisterName.Width))).pValue;
+                var heightPValue = (await Gvcp.GetRegister(nameof(RegisterName.Height))).pValue;
+                GvcpReply widthWriteReply = (await widthPValue.SetValueAsync(width).ConfigureAwait(false)) as GvcpReply;
+                GvcpReply heightWriteReply = (await heightPValue.SetValueAsync(height).ConfigureAwait(false)) as GvcpReply;
                 bool status = (widthWriteReply.Status == GvcpStatus.GEV_STATUS_SUCCESS && heightWriteReply.Status == GvcpStatus.GEV_STATUS_SUCCESS);
                 if (status)
                 {
-                    long newWidth = (long)(await Gvcp.RegistersDictionary[nameof(RegisterName.Width)].pValue.GetValueAsync().ConfigureAwait(false));
-                    long newHeight = (long)(await Gvcp.RegistersDictionary[nameof(RegisterName.Height)].pValue.GetValueAsync().ConfigureAwait(false));
+                    long newWidth = (long)(await widthPValue.GetValueAsync().ConfigureAwait(false));
+                    long newHeight = (long)(await heightPValue.GetValueAsync().ConfigureAwait(false));
 
                     Width = (uint)newWidth;
                     Height = (uint)newHeight;
                 }
 
-                    await Gvcp.LeaveControl().ConfigureAwait(false);                
-                    return status;
+                await Gvcp.LeaveControl().ConfigureAwait(false);
+                return status;
             }
             catch (Exception)
             {
@@ -442,9 +445,11 @@ namespace GigeVision.Core.Models
             {
                 await Gvcp.TakeControl().ConfigureAwait(false);
             }
+            var offsetXRegister = (await Gvcp.GetRegister(nameof(RegisterName.OffsetX))).register;
+            var offsetYRegister = (await Gvcp.GetRegister(nameof(RegisterName.OffsetY))).register;
             string[] registers = new string[2];
-            registers[0] = string.Format("0x{0:X8}", (await Gvcp.RegistersDictionary[nameof(RegisterName.OffsetX)].register.GetAddressAsync()));
-            registers[1] = string.Format("0x{0:X8}", (await Gvcp.RegistersDictionary[nameof(RegisterName.OffsetY)].register.GetAddressAsync()));
+            registers[0] = string.Format("0x{0:X8}", (await offsetXRegister.GetAddressAsync()));
+            registers[1] = string.Format("0x{0:X8}", (await offsetYRegister.GetAddressAsync()));
             uint[] valueToWrite = new uint[] { offsetX, offsetY };
             bool status = (await Gvcp.WriteRegisterAsync(registers, valueToWrite).ConfigureAwait(false)).Status == GvcpStatus.GEV_STATUS_SUCCESS;
             GvcpReply reply = await Gvcp.ReadRegisterAsync(registers).ConfigureAwait(false);
@@ -504,36 +509,24 @@ namespace GigeVision.Core.Models
         /// <returns></returns>
         public async Task<bool> SyncParameters(int syncAttempts = 1)
         {
+
             try
             {
-                if (Gvcp.IsLoadingXml)
+                if (!await Gvcp.ReadXmlFileAsync(IP))
                     return false;
 
-                if (Gvcp.RegistersDictionary?.Count > 0)
-                {
-                    Width = (uint)await Gvcp.RegistersDictionary[nameof(RegisterName.Width)].pValue.GetValueAsync().ConfigureAwait(false);
-                    Height = (uint)await Gvcp.RegistersDictionary[nameof(RegisterName.Height)].pValue.GetValueAsync().ConfigureAwait(false);
-                    OffsetX = (uint)await Gvcp.RegistersDictionary[nameof(RegisterName.OffsetX)].pValue.GetValueAsync().ConfigureAwait(false);
-                    OffsetY = (uint)await Gvcp.RegistersDictionary[nameof(RegisterName.OffsetY)].pValue.GetValueAsync().ConfigureAwait(false);
-                    PixelFormat = (PixelFormat)(uint)await Gvcp.RegistersDictionary[nameof(RegisterName.PixelFormat)].pValue.GetValueAsync().ConfigureAwait(false);
-                    bytesPerPixel = 3;
-                }
-                else
-                {
-                    for (int i = 0; i < syncAttempts; i++)
-                    {
-                        await Gvcp.ReadAllRegisterAddressFromCameraAsync().ConfigureAwait(false);
+                var widthPValue = (await Gvcp.GetRegister(nameof(RegisterName.Width))).pValue;
+                var heightPValue = (await Gvcp.GetRegister(nameof(RegisterName.Height))).pValue;
+                var offsetXPValue = (await Gvcp.GetRegister(nameof(RegisterName.OffsetX))).pValue;
+                var offsetYPValue = (await Gvcp.GetRegister(nameof(RegisterName.OffsetY))).pValue;
+                var pixelFormatPValue = (await Gvcp.GetRegister(nameof(RegisterName.PixelFormat))).pValue;
+                Width = (uint)await widthPValue.GetValueAsync().ConfigureAwait(false);
+                Height = (uint)await heightPValue.GetValueAsync().ConfigureAwait(false);
+                OffsetX = (uint)await offsetXPValue.GetValueAsync().ConfigureAwait(false);
+                OffsetY = (uint)await offsetYPValue.GetValueAsync().ConfigureAwait(false);
+                PixelFormat = (PixelFormat)(uint)await pixelFormatPValue.GetValueAsync().ConfigureAwait(false);
+                bytesPerPixel = 3;
 
-                        if (Gvcp.RegistersDictionary?.Count > 0)
-                        {
-                            await SyncParameters(0);
-                            return true;
-                        }
-
-                    }
-
-                    return false;
-                }
             }
             catch (Exception ex)
             {
@@ -598,7 +591,6 @@ namespace GigeVision.Core.Models
 
         private async void CameraIpChanged(object sender, EventArgs e)
         {
-            Gvcp.RegistersDictionary?.Clear();
             await SyncParameters().ConfigureAwait(false);
         }
     }
