+++ GigeVision.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenICam", "GenICam\GenICam.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GigeVisionLibrary.Test.Wpf", "GigeVisionLibrary.Test.Wpf\GigeVisionLibrary.Test.Wpf.csproj", "{928BED3F-4E80-4973-98CF-A593F708CB0B}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenICam.Tests", "GenICam.Tests\GenICam.Tests.csproj", "{76194E2F-D17E-42F6-8EF0-E084C408A022}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -100,6 +102,24 @@ Global
 		{928BED3F-4E80-4973-98CF-A593F708CB0B}.Release|x64.Build.0 = Release|Any CPU
 		{928BED3F-4E80-4973-98CF-A593F708CB0B}.Release|x86.ActiveCfg = Release|Any CPU
 		{928BED3F-4E80-4973-98CF-A593F708CB0B}.Release|x86.Build.0 = Release|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Debug|x64.Build.0 = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Debug|x86.Build.0 = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.PublishLocally|Any CPU.ActiveCfg = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.PublishLocally|Any CPU.Build.0 = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.PublishLocally|x64.ActiveCfg = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.PublishLocally|x64.Build.0 = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.PublishLocally|x86.ActiveCfg = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.PublishLocally|x86.Build.0 = Debug|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Release|Any CPU.Build.0 = Release|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Release|x64.ActiveCfg = Release|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Release|x64.Build.0 = Release|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Release|x86.ActiveCfg = Release|Any CPU
+		{76194E2F-D17E-42F6-8EF0-E084C408A022}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -109,6 +129,7 @@ Global
 		{6EBD3983-8BB0-4FD4-B1EC-04335D7AE962} = {B120F0DB-D7D6-4ABA-845B-32117EB4B1B2}
 		{329D0921-AC4C-430D-92CD-B387AE869764} = {B120F0DB-D7D6-4ABA-845B-32117EB4B1B2}
 		{928BED3F-4E80-4973-98CF-A593F708CB0B} = {0307491D-B8FC-42AE-8EAB-69D2C072973B}
+		{76194E2F-D17E-42F6-8EF0-E084C408A022} = {0307491D-B8FC-42AE-8EAB-69D2C072973B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C6692468-3D71-4CF6-BBB2-3E938FFB2E6F}
