+++ GigeVisionLibrary.Test.Wpf/GigeVisionLibrary.Test.Wpf.csproj
@@ -9,13 +9,10 @@
    
   <ItemGroup>
     <!--<PackageReference Include="GigeVision.Core" Version="2.1.1" />-->
+    <PackageReference Include="GigeVision.Core" Version="2.3.4" />
     <PackageReference Include="Prism.DryIoc" Version="8.1.97" />
     <PackageReference Include="PropertyChanged.Fody" Version="3.4.1" />
     <PackageReference Include="Stira.ImageLibrary.Wpf" Version="3.2.0" />
     <PackageReference Include="Stira.WpfFlatStyle" Version="1.9.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GigeVision.Core\GigeVision.Core.csproj" />
   </ItemGroup>
 </Project>
